fix: add support for Heltec dev-env 3.0.0 The new version of Heltec ESP32 Development Environment is based on IDF 5

+++ src/SDWrapper/sd_wrapper.cpp
@@ -230,7 +230,7 @@ void SDWrapper::openFile(const char *filename, bool writing) {
 
         // This reading is taken in SDWrapper::openFile.
         if (opened_as_write) {
-            cached_size = max(cached_size, image.size());   // The file may have grown since we opened
+            cached_size = max(cached_size, (uint32_t)image.size());   // The file may have grown since we opened
 
             image.seek(cached_size);
             image.seek(0);  // To be tidy
